fear: add a reuseable component
